add ispyb reading stub to rotation scan plan
